refactor(cli): split evidence schema facade
